fixed crash, added persistence
When encountering a deleted file on Openload, the script will now state as much instead of crashing.
Reloading the page will no longer remove the link display section. It has to either be closed or overwritten by a new batch job.
Behind the scenes there is also a new make.js file to aid in building the userscript.